requirements.txt
blinker==1.9.0
bottle==0.13.4
cffi==1.17.1
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
proxy_tools==0.1.0
pycparser==2.22
pythonnet==3.0.5
pywebview==6.0
qrcode==8.2
typing_extensions==4.14.1
Werkzeug==3.1.3
